first version of reinforcement learning for popout, not debugged

+++ src/popout/search/rlearner/ReinforceLearnerAction.java
@@ -12,7 +12,7 @@ package popout.search.rlearner;
 public class ReinforceLearnerAction {
   private float state_id;   //the state this action was performed in
   private int action_id;  //the action
-  private int reward;     //the immediate reward given for choosing this action
+  private float reward;     //the immediate reward given for choosing this action
   
   //constructors
   public ReinforceLearnerAction(){
@@ -21,7 +21,7 @@ public class ReinforceLearnerAction {
     this.reward = 0;
   }
   
-  public ReinforceLearnerAction(float sid, int aid, int reward){
+  public ReinforceLearnerAction(float sid, int aid, float reward){
     this.state_id = sid;
     this.action_id = aid;
     this.reward = reward;
@@ -30,5 +30,5 @@ public class ReinforceLearnerAction {
   //getters
   public float getStateID(){ return state_id; }
   public int getActionID(){ return action_id; }
-  public int getReward(){ return reward; }
+  public float getReward(){ return reward; }
 }
